+++ tests/Fixtures/Data/PhoneNumberData.php
@@ -1,0 +1,27 @@
+<?php declare(strict_types=1);
+
+/**
+ * Copyright (C) Brian Faust
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Tests\Fixtures\Data;
+
+use Cline\PhoneNumber\PhoneNumber;
+use Cline\Struct\AbstractData;
+use Cline\Struct\Attributes\AsPhoneNumber;
+
+/**
+ * @author Brian Faust <brian@cline.sh>
+ * @psalm-immutable
+ */
+final readonly class PhoneNumberData extends AbstractData
+{
+    public function __construct(
+        public PhoneNumber $international,
+        #[AsPhoneNumber(regionCode: 'US')]
+        public PhoneNumber $local,
+    ) {}
+}